Clean Code: Put Clases/Enums in seperate files.
