+++ MPSSELight/devices/FT2232D.cs
@@ -29,30 +29,14 @@ using System.Threading.Tasks;
 
 namespace MPSSELight
 {
-    [Flags]
-    public enum Ft2232dPin
-    {
-        None = 0,
-
-        ADBUS0 = 1,
-        ADBUS1 = 1 << 1,
-        ADBUS2 = 1 << 2,
-        ADBUS3 = 1 << 3,
-        ADBUS4 = 1 << 4,
-        ADBUS5 = 1 << 5,
-        ADBUS6 = 1 << 6,
-        ADBUS7 = 1 << 7,
-
-        ACBUS0 = 1,
-        ACBUS1 = 1 << 1,
-        ACBUS2 = 1 << 2,
-        ACBUS3 = 1 << 3,
-    }
-
     public class FT2232D : MpsseDevice
     {
-        public FT2232D(String serialNumber) : base(serialNumber) { }
+        public FT2232D(String serialNumber) : base(serialNumber)
+        {
+        }
 
-        public FT2232D(String serialNumber, MpsseParams param) : base(serialNumber, param) { }
+        public FT2232D(String serialNumber, MpsseParams param) : base(serialNumber, param)
+        {
+        }
     }
 }
